+++ client/client.go
@@ -34,6 +34,8 @@ type Client struct {
 
 // New will return a new Prolific client.
 func New() Client {
+	viper.SetDefault("PROLIFIC_URL", "https://api.prolific.co")
+
 	client := Client{
 		Client:  http.DefaultClient,
 		Token:   viper.GetString("PROLIFIC_TOKEN"),