+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## next
 
+## 0.0.10
+
 - Set a default value for `PROLIFIC_URL`. It will default to `https://api.prolific.co`.
 - Ability to get your hook subscriptions via `hook list`.